test(agent-control-plane): split control plane tests
